decoupled db/models from db/crud corresponding to db interaction
